install axios and thunk, refactoring of reducers
