+++ src/api.js
@@ -1,0 +1,31 @@
+import axios from 'axios';
+
+const Error = {
+  UNAUTHORIZED: 401
+};
+
+export const createAPI = (onUnauthorized) => {
+
+  const api = axios.create({
+    baseURL: `https://htmlacademy-react-3.appspot.com/six-cities`,
+    withCredentials: true,
+    timeout: 5000
+  });
+
+  const onSuccess = (response) => {
+    return response;
+  };
+
+  const onFail = (err) => {
+    const {response} = err;
+    if (response.status === Error.UNAUTHORIZED) {
+      onUnauthorized();
+      throw err;
+    }
+    throw err;
+  };
+
+  api.interceptors.response.use(onSuccess, onFail);
+
+  return api;
+};
