+++ src/index.js
@@ -3,6 +3,9 @@ import ReactDOM from 'react-dom';
 import {Provider} from 'react-redux';
 import App from './components/app/app.jsx';
 import {store} from './store';
+import {createAPI} from './api';
+
+// const api = createAPI();
 
 ReactDOM.render(
     <Provider store={store}>
